// Frontend/src/enums/CM/cm004.ts
enum WorkProcessDetailType {
  InProcess = 'inProcess',
  Complete = 'Complete',
}

enum CmDisbursementApprovalStatus {
  Draft = 'Draft',
  Assigned = 'Assigned',
  WaitingApproval = 'WaitingApproval',
  Approved = 'Approved',
  Rejected = 'Rejected',
}

export {
  WorkProcessDetailType,
  CmDisbursementApprovalStatus,
}